添加treeDp

+++ src/main/java/com/asiainfo/security/controller/RoleMongoController.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
 /**
@@ -17,7 +18,7 @@ import org.springframework.web.bind.annotation.*;
  * @version 2019/10/1213:56
  */
 @RestController
-//@RequestMapping("api")
+@RequestMapping("admin")
 public class RoleMongoController {
     @Autowired
     private RoleMongoService roleMongoService;
@@ -36,12 +37,16 @@ public class RoleMongoController {
 //    ("新增角色")
     @PostMapping(value = "/roles")
     public ResponseEntity create(@RequestBody RoleDP resources){
+        if (resources == null|| StringUtils.isEmpty(resources.getRoleName()))
+            throw new RuntimeException("角色名不能为空...");
         return new ResponseEntity(roleMongoService.create(resources),HttpStatus.CREATED);
     }
 
 //    ("修改角色")
     @PutMapping(value = "/roles")
     public ResponseEntity update(@RequestBody RoleDP resources){
+        if (resources == null||StringUtils.isEmpty(resources.getRoleName()))
+            throw new RuntimeException("角色名不能为空...");
         roleMongoService.update(resources);
         return new ResponseEntity(HttpStatus.OK);
     }
@@ -49,15 +54,16 @@ public class RoleMongoController {
 //    ("删除角色")
     @DeleteMapping(value = "/roles/{id}")
     public ResponseEntity delete(@PathVariable String id){
+        if (StringUtils.isEmpty(id))throw new RuntimeException("不存在该角色id值...");
         roleMongoService.delete(id);
         return new ResponseEntity(HttpStatus.OK);
     }
 
 
 
-
-
-
-
-
+    //    ("删除用户")
+    @GetMapping(value = "/roles/tree")
+    public ResponseEntity createTree(RoleMongoCriteria criteria){
+        return new ResponseEntity(roleMongoService.createTree(criteria),HttpStatus.OK);
+    }
 }
